text

+++ stock/scheme.php
@@ -2,31 +2,33 @@
 
 include_once $_SERVER["DOCUMENT_ROOT"] . "/db/db.php";
 
-query("DROP TABLE IF EXISTS `users`;");
-query("CREATE TABLE IF NOT EXISTS `users` (
+$drop = get_int("drop", 0);
+if ($drop == 1) {
+    query("DROP TABLE IF EXISTS `users`;");
+    query("CREATE TABLE IF NOT EXISTS `users` (
     `user_id` int(11) NOT NULL AUTO_INCREMENT,
     `token` varchar(128) COLLATE utf32_bin NOT NULL,
     PRIMARY KEY (`user_id`)
     ) ENGINE=MyISAM DEFAULT CHARSET=utf32 COLLATE=utf32_bin;");
 
-query("DROP TABLE IF EXISTS `balances`;");
-query("CREATE TABLE IF NOT EXISTS `balances` (
+    query("DROP TABLE IF EXISTS `balances`;");
+    query("CREATE TABLE IF NOT EXISTS `balances` (
     `user_id` int(11) NOT NULL,
     `ticker` varchar(10) COLLATE utf32_bin NOT NULL,
     `spot` double NOT NULL,
     `blocked` double NOT NULL
     ) ENGINE=MyISAM DEFAULT CHARSET=utf32 COLLATE=utf32_bin;");
 
-query("DROP TABLE IF EXISTS `currencies`;");
-query("CREATE TABLE IF NOT EXISTS `currencies` (
+    query("DROP TABLE IF EXISTS `currencies`;");
+    query("CREATE TABLE IF NOT EXISTS `currencies` (
   `ticker` varchar(10) COLLATE utf32_bin NOT NULL,
   `rate` double NOT NULL,
-  PRIMARY KEY (`currency_id`)
+  PRIMARY KEY (`ticker`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf32 COLLATE=utf32_bin;");
 
-query("DROP TABLE IF EXISTS `orders`;");
-query("CREATE TABLE IF NOT EXISTS `orders` (
-  `order_id` int(11) NOT NULL,
+    query("DROP TABLE IF EXISTS `orders`;");
+    query("CREATE TABLE IF NOT EXISTS `orders` (
+  `order_id` int(11) NOT NULL AUTO_INCREMENT,
   `ticker` varchar(10) COLLATE utf32_bin NOT NULL,
   `user_id` int(11) NOT NULL,
   `is_sell` int(11) NOT NULL,
@@ -34,5 +36,12 @@ query("CREATE TABLE IF NOT EXISTS `orders` (
   `price` double NOT NULL,
   `amount` double NOT NULL,
   `filled` double DEFAULT 0,
-  `timestamp` int(11) NOT NULL
+  `timestamp` int(11) NOT NULL,
+   PRIMARY KEY (`order_id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf32 COLLATE=utf32_bin;");
+} else {
+    query("DELETE FROM `users`;");
+    query("DELETE FROM `balances`;");
+    query("DELETE FROM `currencies`;");
+    query("DELETE FROM `orders`;");
+}
